Put API class into the Network module
