+++ Client/Source/KiwiApp_Network/KiwiApp_DocumentBrowser.h
@@ -26,12 +26,14 @@
 #include <juce_data_structures/juce_data_structures.h>
 #include <juce_events/juce_events.h>
 
-#include "KiwiApp_Api.h"
+#include <KiwiNetwork/KiwiNetwork_Api.h>
 
 #include "../KiwiApp_General/KiwiApp_StoredSettings.h"
 
 namespace kiwi
 {
+    using network::Api;
+    
     // ================================================================================ //
     //                                 DOCUMENT BROWSER                                 //
     // ================================================================================ //
